+++ include/BohdanAna.h
@@ -27,31 +27,31 @@ class BohdanAna : public marlin::Processor, EventDisplayer {
         void resetVariables();
     private:
         int _nEvent{};
-        double _bField{};
+        float _bField{};
         
         std::unique_ptr<TFile> _file;
         std::unique_ptr<TTree> _tree;
 
         int _pdg{};
-        double _dEdx{};
-        double _omegaIP{};
-        double _omegaECAL{};
-        double _tanLambdaIP{};
-        double _tanLambdaECAL{};
+        float _dEdx{};
+        float _omegaIP{};
+        float _omegaECAL{};
+        float _tanLambdaIP{};
+        float _tanLambdaECAL{};
 
-        std::array<double, 3> _mcMom{};
-        std::array<double, 3> _recoIpMom{};
-        std::array<double, 3> _recoCaloPos{};
-        std::array<double, 3> _recoCaloMom{};
+        std::array<float, 3> _mcMom{};
+        std::array<float, 3> _recoIpMom{};
+        std::array<float, 3> _recoCaloPos{};
+        std::array<float, 3> _recoCaloMom{};
 
 
-        double _trackLength_IDR{};
-        double _trackLength_SHA_phiLambda_IP{};
-        double _trackLength_SHA_phiZed_IP{};
-        double _trackLength_SHA_zedLambda_IP{};
-        double _trackLength_SHA_phiLambda_ECAL{};
-        double _trackLength_SHA_phiZed_ECAL{};
-        double _trackLength_SHA_zedLambda_ECAL{};
+        float _trackLength_IDR{};
+        float _trackLength_SHA_phiLambda_IP{};
+        float _trackLength_SHA_phiZed_IP{};
+        float _trackLength_SHA_zedLambda_IP{};
+        float _trackLength_SHA_phiLambda_ECAL{};
+        float _trackLength_SHA_phiZed_ECAL{};
+        float _trackLength_SHA_zedLambda_ECAL{};
         TrackLengthResult _trackLength_IKF_phiLambda;
         TrackLengthResult _trackLength_IKF_phiZed;
         TrackLengthResult _trackLength_IKF_zedLambda;
@@ -63,19 +63,19 @@ class BohdanAna : public marlin::Processor, EventDisplayer {
         int _layoutClosest = -1;
         int _layerClosest = -1;
         bool _cleanClosestHit = false;
-        std::vector<double> _resolutions = {0, 1, 5, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 200, 300};
-        std::array<double, 15> _tofClosest{};
-        std::array<double, 15> _tofAverage{};
-        std::array<double, 15> _tofSET{};
-        std::array<double, 15> _tofFit{};
+        std::vector<float> _resolutions = {0, 1, 5, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 200, 300};
+        std::array<float, 15> _tofClosest{};
+        std::array<float, 15> _tofAverage{};
+        std::array<float, 15> _tofSET{};
+        std::array<float, 15> _tofFit{};
 
         int _nHits;
-        std::vector<double> _xHit;
-        std::vector<double> _yHit;
-        std::vector<double> _zHit;
-        std::vector<double> _tHit;
+        std::vector<float> _xHit;
+        std::vector<float> _yHit;
+        std::vector<float> _zHit;
+        std::vector<float> _tHit;
         std::vector<int> _layerHit;
-        std::vector<double> _energyHit;
+        std::vector<float> _energyHit;
 
         MarlinTrk::IMarlinTrkSystem* _trkSystem = nullptr;
         TApplication _application = TApplication("app", 0, nullptr);
