+++ include/TOF.h
@@ -15,20 +15,20 @@ std::vector<EVENT::CalorimeterHit*> selectFrankEcalHits( EVENT::Cluster* cluster
 
 EVENT::CalorimeterHit* getClosestHit( EVENT::Cluster* cluster, const dd4hep::rec::Vector3D& posAtEcal );
 
-double getHitTof( EVENT::CalorimeterHit* hit, const dd4hep::rec::Vector3D& posAtEcal, double timeResolution) ;
+float getHitTof( EVENT::CalorimeterHit* hit, const dd4hep::rec::Vector3D& posAtEcal, float timeResolution) ;
 int getHitCaloType( EVENT::CalorimeterHit* hit );
 int getHitCaloID( EVENT::CalorimeterHit* hit );
 int getHitCaloLayout( EVENT::CalorimeterHit* hit );
 int getHitCaloLayer( EVENT::CalorimeterHit* hit );
 EVENT::MCParticle* getHitEarliestMC( EVENT::CalorimeterHit* hit, const UTIL::LCRelationNavigator& navToSimCalorimeterHits );
 
-double getTofFrankAvg( const std::vector<EVENT::CalorimeterHit*>& selectedHits, const dd4hep::rec::Vector3D& posAtEcal, double timeResolution );
+float getTofFrankAvg( const std::vector<EVENT::CalorimeterHit*>& selectedHits, const dd4hep::rec::Vector3D& posAtEcal, float timeResolution );
 
-double getTofFrankFit( const std::vector<EVENT::CalorimeterHit*>& selectedHits, const dd4hep::rec::Vector3D& posAtEcal, double timeResolution );
+float getTofFrankFit( const std::vector<EVENT::CalorimeterHit*>& selectedHits, const dd4hep::rec::Vector3D& posAtEcal, float timeResolution );
 
-double getTofSET(EVENT::Track* track, double timeResolution );
+float getTofSET(EVENT::Track* track, float timeResolution );
 
-double getTofPhotonTrue(EVENT::MCParticle* mc );
+float getTofPhotonTrue(EVENT::MCParticle* mc );
 
 
 #endif
